Split SpecialFish headers
